ingredient area splited to few components
