+++ src/components/ingredient-area/ingredient-area.module.css
@@ -1,0 +1,7 @@
+.ingredientsGrid {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  row-gap: 32px;
+  column-gap: 24px;
+  margin: auto;
+}
