+++ src/components/ingredient-item/ingredient-item.module.css
@@ -1,0 +1,26 @@
+.ingredientsCard {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  position: relative;
+  width: 272px;
+  cursor: pointer;
+}
+
+
+
+.ingredientName {
+  text-align: center;
+}
+
+.priceLabel {
+  display: flex;
+  gap: 8px;
+  align-items: center;
+}
+
+.counter {
+  position: absolute;
+  top: 0;
+  right: 0;
+}
