+++ src/components/nav-ingredient/nav-ingredient.module.css
@@ -1,0 +1,8 @@
+.navIngredient {
+  display: flex;
+}
+
+.navIngredientItem {
+  text-decoration: none;
+  scroll-behavior: smooth;
+}
